Moved menu and actionbar code to separate class
