+++ DroneControllerApp/app/src/main/res/values/strings.xml
@@ -78,4 +78,5 @@
     <string name="title_activity_mission">MissionActivity</string>
     <string name="title_activity_create_mission">CreateMissionActivity</string>
     <string name="title_activity_confirm_mission">ConfirmMissionActivity</string>
+    <string name="title_activity_menu">MenuActivity</string>
 </resources>
